# Perf improvement: line's envelope method