## Description

Improves the performance of line's envelope method. This is a (smallish...) bottleneck in Intersection/Union etc.

## Check List

Have you:

- Added unit tests? N/A, existing.

- Add cmprefimpl tests? (if appropriate?) N/A, existing.

## Related Issue

- N/A

## Benchmark Results

Results are significant for the actual envelope method. I wasn't able to time the impact on Intersection, since the variance is higher than the actual saving. I calculated the saving to be around ~3% though (so the gain is marginal). I figure 3% since the `envelope` method shows up as 3% in the pprof profile, and as can be seen in the benchmarks below, this perf change basically makes the method call free.

```
COMPARISON
name              old time/op    new time/op    delta
LineEnvelope/0-4    19.4ns ±15%     1.6ns ± 7%  -91.67%  (p=0.000 n=14+13)
LineEnvelope/1-4    18.4ns ± 5%     1.1ns ± 9%  -94.28%  (p=0.000 n=12+13)
LineEnvelope/2-4    18.8ns ±17%     1.3ns ±10%  -92.87%  (p=0.000 n=15+14)
LineEnvelope/3-4    18.3ns ±12%     1.2ns ±14%  -93.40%  (p=0.000 n=15+15)

name              old alloc/op   new alloc/op   delta
LineEnvelope/0-4     0.00B          0.00B          ~     (all equal)
LineEnvelope/1-4     0.00B          0.00B          ~     (all equal)
LineEnvelope/2-4     0.00B          0.00B          ~     (all equal)
LineEnvelope/3-4     0.00B          0.00B          ~     (all equal)

name              old allocs/op  new allocs/op  delta
LineEnvelope/0-4      0.00           0.00          ~     (all equal)
LineEnvelope/1-4      0.00           0.00          ~     (all equal)
LineEnvelope/2-4      0.00           0.00          ~     (all equal)
LineEnvelope/3-4      0.00           0.00          ~     (all equal)
```